initial playwright config
