+++ .gitignore
@@ -34,3 +34,8 @@ dist/**/*
 
 # ignore yarn.lock
 yarn.lock
+
+# Ignore playwright generated files
+/playwright/results/
+/playwright/report/
+/playwright/.cache/
