+++ package.json
@@ -25,6 +25,7 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@playwright/test": "^1.49.0",
     "@shopify/eslint-plugin": "^46.0.0",
     "@types/cookie-parser": "^1.4.7",
     "@types/express": "^4.17.21",
