+++ playwright/.auth/admin.json
@@ -1,0 +1,15 @@
+{
+  "cookies": [
+    {
+      "name": "jwt",
+      "value": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjoiMDQ0ZDk0YzUtOTFiYS00OTVlLWE3MTgtMzFjNTk3YTBhMzBiIiwibmFtZSI6IlRvbSBBZG1pbiIsImVtYWlsIjoidG9tLmFkbWluQGV4YW1wbGUuY29tIn0sImlhdCI6MTczMjExNzczOX0.VzdAB9qbQOtuyB4xPQvreIWLSku8d4xsUZRmIRBtmcU",
+      "domain": "localhost",
+      "path": "/",
+      "expires": -1,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    }
+  ],
+  "origins": []
+}
