+++ playwright/.auth/approver.json
@@ -1,0 +1,15 @@
+{
+  "cookies": [
+    {
+      "name": "jwt",
+      "value": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjoiY2UwODcyN2UtZGQzZi00OGNjLTkyMWEtY2FlNWM0ZGQ0YTE4IiwibmFtZSI6IkZyYW5rIEFwcHJvdmVyIiwiZW1haWwiOiJmcmFuay5hcHByb3ZlckBleGFtcGxlLmNvbSJ9LCJpYXQiOjE3MzIxMTc3Mzl9.vfpIpkvxbB_mU-ZMk-zwKHMX7HV6NSEE6WXx-3TDSAU",
+      "domain": "localhost",
+      "path": "/",
+      "expires": -1,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    }
+  ],
+  "origins": []
+}
