+++ playwright/.auth/contexts.ts
@@ -1,0 +1,5 @@
+import path from 'node:path';
+
+export const adminContext = path.join(__dirname, './admin.json');
+export const publisherContext = path.join(__dirname, './publisher.json');
+export const approverContext = path.join(__dirname, './approver.json');
