+++ src/interfaces/user.interface.ts
@@ -1,6 +1,7 @@
 export interface User {
     id: string;
     email: string;
-    firstName: string;
-    lastName: string;
+    name: string;
+    givenName?: string;
+    familyName?: string;
 }
