+++ src/views/partials/header.ejs
@@ -83,11 +83,11 @@
                         <% } %>
                         <li class="govuk-header__navigation-item govuk-header__navigation-item--active">
                             <% if (i18n.language === 'en-GB') { %>
-                                <a class="govuk-header__link" href="?lang=cy-GB">
+                                <a class="govuk-header__link" lang="cy" hreflang="cy" href="?lang=cy-GB">
                                     Cymraeg
                                 </a>
                             <% } else { %>
-                                <a class="govuk-header__link" href="?lang=en-GB">
+                                <a class="govuk-header__link" lang="en" hreflang="en" href="?lang=en-GB">
                                     English
                                 </a>
                             <% } %>
